refactor: 디렉토리 복잡도를 낮추기 위해 stories 파일 분리
