+++ src/stories/Card.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react';
-import Card from './Card';
+import Card from '../components/Card/Card';
 
 const meta = {
   component: Card,
