+++ src/stories/CardOwnerName.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react';
-import CardOwnerName from './CardOwnerName';
+import CardOwnerName from '../components/CardOwnerName/CardOwnerName';
 
 const meta = {
   component: CardOwnerName,
